OK association Client

+++ src/projects/projects.service.ts
@@ -1,15 +1,17 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateProjectDto } from './dto/create-project.dto';
 import { UpdateProjectDto } from './dto/update-project.dto';
 import { Project } from './entities/project.entity';
+import { Dev } from '../devs/entities/dev.entity';
+import { Client } from '../clients/entities/client.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
 @Injectable()
 export class ProjectsService {
   constructor(
-    @InjectRepository(Project) 
-    private projectRepository: Repository<Project>
+    @InjectRepository(Project)
+    private projectRepository: Repository<Project>,
   ) {}
 
   create(createProjectDto) {
@@ -18,11 +20,11 @@ export class ProjectsService {
   }
 
   findAll() {
-    this.projectRepository.find();
+    return this.projectRepository.find();
   }
 
   findOne(id) {
-    this.projectRepository.findOne(id);
+    return this.projectRepository.findOne(id);
   }
 
   update(id, updateProjectDto) {
@@ -30,7 +32,24 @@ export class ProjectsService {
     return this.projectRepository.findOne(id);
   }
 
-  remove(id) {
-    this.projectRepository.delete(id);
+  async remove(id): Promise<string> {
+    const project = this.projectRepository.findOne(id);
+    console.log(project);
+    if (!project) {
+      throw new NotFoundException(`Project with id ${id} not found`);
+    } else {
+      this.projectRepository.delete(id);
+      return `Project with id ${id} has been deleted`;
+    }
+  }
+
+  async findClientByProjectId(projectId: number): Promise<Client> {
+    const project = await this.projectRepository.findOne({
+      where: {id: projectId},
+      relations: ['client'] });
+    if (!project) {
+      throw new NotFoundException(`Project with id ${projectId} not found`);
+    }
+    return project.client;
   }
 }
